@@ -1,8 +1,9 @@
 module github.com/beego/beego/v2
 
-go 1.18
+go 1.20
 
 require (
+	github.com/DATA-DOG/go-sqlmock v1.5.1
 	github.com/beego/x2j v0.0.0-20131220205130-a0352aadc542
 	github.com/bits-and-blooms/bloom/v3 v3.6.0
 	github.com/bradfitz/gomemcache v0.0.0-20190913173617-a41fca850d0b
