@@ -42,7 +42,7 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"path"
+	"path/filepath"
 	"strings"
 	"time"
 
@@ -73,7 +73,6 @@ func NewBeegoRequestWithCtx(ctx context.Context, rawurl, method string) *BeegoHT
 	if err != nil {
 		logs.Error("%+v", berror.Wrapf(err, InvalidURLOrMethod, "invalid raw url or method: %s %s", rawurl, method))
 	}
-
 	return &BeegoHTTPRequest{
 		url:     rawurl,
 		req:     req,
@@ -81,6 +80,9 @@ func NewBeegoRequestWithCtx(ctx context.Context, rawurl, method string) *BeegoHT
 		files:   map[string]string{},
 		setting: defaultSetting,
 		resp:    &http.Response{},
+		copyBody: func() io.ReadCloser {
+			return nil
+		},
 	}
 }
 
@@ -117,7 +119,10 @@ type BeegoHTTPRequest struct {
 	files   map[string]string
 	setting BeegoHTTPSettings
 	resp    *http.Response
-	body    []byte
+	// body the response body, not the request body
+	body []byte
+	// copyBody support retry strategy to avoid copy request body
+	copyBody func() io.ReadCloser
 }
 
 // GetRequest returns the request object
@@ -281,21 +286,24 @@ func (b *BeegoHTTPRequest) PostFile(formname, filename string) *BeegoHTTPRequest
 func (b *BeegoHTTPRequest) Body(data interface{}) *BeegoHTTPRequest {
 	switch t := data.(type) {
 	case string:
-		bf := bytes.NewBufferString(t)
-		b.req.Body = io.NopCloser(bf)
-		b.req.GetBody = func() (io.ReadCloser, error) {
-			return io.NopCloser(bf), nil
-		}
-		b.req.ContentLength = int64(len(t))
+		b.reqBody([]byte(t))
 	case []byte:
-		bf := bytes.NewBuffer(t)
-		b.req.Body = io.NopCloser(bf)
-		b.req.GetBody = func() (io.ReadCloser, error) {
-			return io.NopCloser(bf), nil
-		}
-		b.req.ContentLength = int64(len(t))
+		b.reqBody(t)
 	default:
 		logs.Error("%+v", berror.Errorf(UnsupportedBodyType, "unsupported body data type: %s", t))
+	}
+	return b
+}
+
+func (b *BeegoHTTPRequest) reqBody(data []byte) *BeegoHTTPRequest {
+	body := io.NopCloser(bytes.NewReader(data))
+	b.req.Body = body
+	b.req.GetBody = func() (io.ReadCloser, error) {
+		return body, nil
+	}
+	b.req.ContentLength = int64(len(data))
+	b.copyBody = func() io.ReadCloser {
+		return io.NopCloser(bytes.NewReader(data))
 	}
 	return b
 }
@@ -307,11 +315,7 @@ func (b *BeegoHTTPRequest) XMLBody(obj interface{}) (*BeegoHTTPRequest, error) {
 		if err != nil {
 			return b, berror.Wrap(err, InvalidXMLBody, "obj could not be converted to XML data")
 		}
-		b.req.Body = io.NopCloser(bytes.NewReader(byts))
-		b.req.GetBody = func() (io.ReadCloser, error) {
-			return io.NopCloser(bytes.NewReader(byts)), nil
-		}
-		b.req.ContentLength = int64(len(byts))
+		b.reqBody(byts)
 		b.req.Header.Set(contentTypeKey, "application/xml")
 	}
 	return b, nil
@@ -324,8 +328,7 @@ func (b *BeegoHTTPRequest) YAMLBody(obj interface{}) (*BeegoHTTPRequest, error) 
 		if err != nil {
 			return b, berror.Wrap(err, InvalidYAMLBody, "obj could not be converted to YAML data")
 		}
-		b.req.Body = io.NopCloser(bytes.NewReader(byts))
-		b.req.ContentLength = int64(len(byts))
+		b.reqBody(byts)
 		b.req.Header.Set(contentTypeKey, "application/x+yaml")
 	}
 	return b, nil
@@ -338,8 +341,7 @@ func (b *BeegoHTTPRequest) JSONBody(obj interface{}) (*BeegoHTTPRequest, error) 
 		if err != nil {
 			return b, berror.Wrap(err, InvalidJSONBody, "obj could not be converted to JSON body")
 		}
-		b.req.Body = io.NopCloser(bytes.NewReader(byts))
-		b.req.ContentLength = int64(len(byts))
+		b.reqBody(byts)
 		b.req.Header.Set(contentTypeKey, "application/json")
 	}
 	return b, nil
@@ -493,7 +495,7 @@ func (b *BeegoHTTPRequest) doRequest(_ context.Context) (*http.Response, error) 
 func (b *BeegoHTTPRequest) sendRequest(client *http.Client) (resp *http.Response, err error) {
 	// retries default value is 0, it will run once.
 	// retries equal to -1, it will run forever until success
-	// retries is setted, it will retries fixed times.
+	// retries is set, it will retry fixed times.
 	// Sleeps for a 400ms between calls to reduce spam
 	for i := 0; b.setting.Retries == -1 || i <= b.setting.Retries; i++ {
 		resp, err = client.Do(b.req)
@@ -501,6 +503,7 @@ func (b *BeegoHTTPRequest) sendRequest(client *http.Client) (resp *http.Response
 			return
 		}
 		time.Sleep(b.setting.RetryDelay)
+		b.req.Body = b.copyBody()
 	}
 	return nil, berror.Wrap(err, SendRequestFailed, "sending request fail")
 }
@@ -623,7 +626,7 @@ func (b *BeegoHTTPRequest) ToFile(filename string) error {
 
 // Check if the file directory exists. If it doesn't then it's created
 func pathExistAndMkdir(filename string) (err error) {
-	filename = path.Dir(filename)
+	filename = filepath.Dir(filename)
 	_, err = os.Stat(filename)
 	if err == nil {
 		return nil
