@@ -20,8 +20,6 @@ import (
 	"reflect"
 	"strings"
 
-	"github.com/beego/beego/v2/client/orm/internal/logs"
-
 	"github.com/beego/beego/v2/client/orm/internal/models"
 )
 
@@ -163,7 +161,7 @@ func (d *dbBaseMysql) InsertOrUpdate(ctx context.Context, q dbQuerier, mi *model
 		if err == nil {
 			lastInsertId, err := res.LastInsertId()
 			if err != nil {
-				logs.DebugLog.Println(ErrLastInsertIdUnavailable, ':', err)
+				DebugLog.Println(ErrLastInsertIdUnavailable, ':', err)
 				return lastInsertId, ErrLastInsertIdUnavailable
 			} else {
 				return lastInsertId, nil
