@@ -918,7 +918,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 		db   *dbBase
 
 		tCols []string
-		qs    *querySet
+		qs    querySet
 
 		wantRes  string
 		wantArgs []interface{}
@@ -929,7 +929,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBaseMysql(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -955,7 +955,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBaseMysql(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -982,7 +982,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBaseMysql(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1009,7 +1009,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBaseMysql(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1037,7 +1037,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBaseMysql(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1064,7 +1064,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBasePostgres(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1088,7 +1088,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBasePostgres(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1113,7 +1113,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBasePostgres(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1138,7 +1138,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBasePostgres(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1164,7 +1164,7 @@ func TestDbBase_readBatchSQL(t *testing.T) {
 				ins: newdbBasePostgres(),
 			},
 			tCols: []string{"name", "score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1221,7 +1221,7 @@ func TestDbBase_readValuesSQL(t *testing.T) {
 		db   *dbBase
 
 		cols []string
-		qs   *querySet
+		qs   querySet
 
 		wantRes  string
 		wantArgs []interface{}
@@ -1232,7 +1232,7 @@ func TestDbBase_readValuesSQL(t *testing.T) {
 				ins: newdbBaseMysql(),
 			},
 			cols: []string{"T0.`name` name", "T0.`age` age", "T0.`score` score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1256,7 +1256,7 @@ func TestDbBase_readValuesSQL(t *testing.T) {
 				ins: newdbBaseMysql(),
 			},
 			cols: []string{"T0.`name` name", "T0.`age` age", "T0.`score` score"},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1281,7 +1281,7 @@ func TestDbBase_readValuesSQL(t *testing.T) {
 				ins: newdbBasePostgres(),
 			},
 			cols: []string{`T0."name" name`, `T0."age" age`, `T0."score" score`},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1303,7 +1303,7 @@ func TestDbBase_readValuesSQL(t *testing.T) {
 				ins: newdbBasePostgres(),
 			},
 			cols: []string{`T0."name" name`, `T0."age" age`, `T0."score" score`},
-			qs: &querySet{
+			qs: querySet{
 				mi:     mi,
 				cond:   cond,
 				limit:  10,
@@ -1356,7 +1356,7 @@ func TestDbBase_countSQL(t *testing.T) {
 		name string
 		db   *dbBase
 
-		qs *querySet
+		qs querySet
 
 		wantRes  string
 		wantArgs []interface{}
@@ -1366,7 +1366,7 @@ func TestDbBase_countSQL(t *testing.T) {
 			db: &dbBase{
 				ins: newdbBaseMysql(),
 			},
-			qs: &querySet{
+			qs: querySet{
 				mi:       mi,
 				cond:     cond,
 				useIndex: 1,
@@ -1382,7 +1382,7 @@ func TestDbBase_countSQL(t *testing.T) {
 			db: &dbBase{
 				ins: newdbBaseMysql(),
 			},
-			qs: &querySet{
+			qs: querySet{
 				mi:       mi,
 				cond:     cond,
 				useIndex: 1,
@@ -1399,7 +1399,7 @@ func TestDbBase_countSQL(t *testing.T) {
 			db: &dbBase{
 				ins: newdbBasePostgres(),
 			},
-			qs: &querySet{
+			qs: querySet{
 				mi:       mi,
 				cond:     cond,
 				related:  make([]string, 0),
@@ -1413,7 +1413,7 @@ func TestDbBase_countSQL(t *testing.T) {
 			db: &dbBase{
 				ins: newdbBasePostgres(),
 			},
-			qs: &querySet{
+			qs: querySet{
 				mi:       mi,
 				cond:     cond,
 				related:  make([]string, 0),
