@@ -7,7 +7,7 @@ const (
 	// OffsetNewest is the log head offset, i.e. the offset that will be
 	// assigned to the next message that will be produced to the shard.
 	OffsetNewest = "end"
-	// OffsetOldest is the the oldest offset available on the logstore for a
+	// OffsetOldest is the oldest offset available on the logstore for a
 	// shard.
 	OffsetOldest = "begin"
 )
