@@ -19,8 +19,6 @@ import (
 	"fmt"
 	"io"
 	"net"
-
-	"github.com/pkg/errors"
 )
 
 // connWriter implements LoggerInterface.
@@ -56,7 +54,7 @@ func (c *connWriter) Init(config string) error {
 	if res == nil && len(c.Formatter) > 0 {
 		fmtr, ok := GetFormatter(c.Formatter)
 		if !ok {
-			return errors.New(fmt.Sprintf("the formatter with name: %s not found", c.Formatter))
+			return fmt.Errorf("the formatter with name: %s not found", c.Formatter)
 		}
 		c.formatter = fmtr
 	}
