@@ -21,8 +21,6 @@ import (
 	"net"
 	"net/smtp"
 	"strings"
-
-	"github.com/pkg/errors"
 )
 
 // SMTPWriter implements LoggerInterface and is used to send emails via given SMTP-server.
@@ -34,13 +32,16 @@ type SMTPWriter struct {
 	FromAddress        string   `json:"fromAddress"`
 	RecipientAddresses []string `json:"sendTos"`
 	Level              int      `json:"level"`
-	formatter          LogFormatter
-	Formatter          string `json:"formatter"`
+	// InsecureSkipVerify default value: true
+	InsecureSkipVerify bool `json:"insecureSkipVerify"`
+
+	formatter LogFormatter
+	Formatter string `json:"formatter"`
 }
 
 // NewSMTPWriter creates the smtp writer.
 func newSMTPWriter() Logger {
-	res := &SMTPWriter{Level: LevelTrace}
+	res := &SMTPWriter{Level: LevelTrace, InsecureSkipVerify: true}
 	res.formatter = res
 	return res
 }
@@ -48,21 +49,22 @@ func newSMTPWriter() Logger {
 // Init smtp writer with json config.
 // config like:
 //
-//	{
-//		"username":"example@gmail.com",
-//		"password:"password",
-//		"host":"smtp.gmail.com:465",
-//		"subject":"email title",
-//		"fromAddress":"from@example.com",
-//		"sendTos":["email1","email2"],
-//		"level":LevelError
-//	}
+//		{
+//			"username":"example@gmail.com",
+//			"password:"password",
+//			"host":"smtp.gmail.com:465",
+//			"subject":"email title",
+//			"fromAddress":"from@example.com",
+//			"sendTos":["email1","email2"],
+//			"level":LevelError,
+//	     	"insecureSkipVerify": false
+//		}
 func (s *SMTPWriter) Init(config string) error {
 	res := json.Unmarshal([]byte(config), s)
 	if res == nil && len(s.Formatter) > 0 {
 		fmtr, ok := GetFormatter(s.Formatter)
 		if !ok {
-			return errors.New(fmt.Sprintf("the formatter with name: %s not found", s.Formatter))
+			return fmt.Errorf("the formatter with name: %s not found", s.Formatter)
 		}
 		s.formatter = fmtr
 	}
@@ -93,7 +95,7 @@ func (s *SMTPWriter) sendMail(hostAddressWithPort string, auth smtp.Auth, fromAd
 
 	host, _, _ := net.SplitHostPort(hostAddressWithPort)
 	tlsConn := &tls.Config{
-		InsecureSkipVerify: true,
+		InsecureSkipVerify: s.InsecureSkipVerify,
 		ServerName:         host,
 	}
 	if err = client.StartTLS(tlsConn); err != nil {
