@@ -14,9 +14,9 @@
 
 // Package redis for session provider
 //
-// depend on github.com/go-redis/redis
+// depend on github.com/redis/go-redis
 //
-// go install github.com/go-redis/redis
+// go install github.com/redis/go-redis
 //
 // Usage:
 // import(
@@ -41,7 +41,7 @@ import (
 	"sync"
 	"time"
 
-	rediss "github.com/go-redis/redis/v7"
+	rediss "github.com/redis/go-redis/v9"
 
 	"github.com/beego/beego/v2/server/web/session"
 )
@@ -109,7 +109,7 @@ func (rs *SessionStore) SessionRelease(ctx context.Context, w http.ResponseWrite
 		return
 	}
 	c := rs.p
-	c.Set(rs.sid, string(b), time.Duration(rs.maxlifetime)*time.Second)
+	c.Set(ctx, rs.sid, string(b), time.Duration(rs.maxlifetime)*time.Second)
 }
 
 // Provider redis_cluster session provider
@@ -156,14 +156,13 @@ func (rp *Provider) SessionInit(ctx context.Context, maxlifetime int64, cfgStr s
 	}
 
 	rp.poollist = rediss.NewClusterClient(&rediss.ClusterOptions{
-		Addrs:              strings.Split(rp.SavePath, ";"),
-		Password:           rp.Password,
-		PoolSize:           rp.Poolsize,
-		IdleTimeout:        rp.idleTimeout,
-		IdleCheckFrequency: rp.idleCheckFrequency,
-		MaxRetries:         rp.MaxRetries,
+		Addrs:           strings.Split(rp.SavePath, ";"),
+		Password:        rp.Password,
+		PoolSize:        rp.Poolsize,
+		ConnMaxIdleTime: rp.idleTimeout,
+		MaxRetries:      rp.MaxRetries,
 	})
-	return rp.poollist.Ping().Err()
+	return rp.poollist.Ping(ctx).Err()
 }
 
 // for v1.x
@@ -218,7 +217,7 @@ func (rp *Provider) initOldStyle(savePath string) {
 // SessionRead read redis_cluster session by sid
 func (rp *Provider) SessionRead(ctx context.Context, sid string) (session.Store, error) {
 	var kv map[interface{}]interface{}
-	kvs, err := rp.poollist.Get(sid).Result()
+	kvs, err := rp.poollist.Get(ctx, sid).Result()
 	if err != nil && err != rediss.Nil {
 		return nil, err
 	}
@@ -237,7 +236,7 @@ func (rp *Provider) SessionRead(ctx context.Context, sid string) (session.Store,
 // SessionExist check redis_cluster session exist by sid
 func (rp *Provider) SessionExist(ctx context.Context, sid string) (bool, error) {
 	c := rp.poollist
-	if existed, err := c.Exists(sid).Result(); err != nil || existed == 0 {
+	if existed, err := c.Exists(ctx, sid).Result(); err != nil || existed == 0 {
 		return false, err
 	}
 	return true, nil
@@ -247,22 +246,22 @@ func (rp *Provider) SessionExist(ctx context.Context, sid string) (bool, error) 
 func (rp *Provider) SessionRegenerate(ctx context.Context, oldsid, sid string) (session.Store, error) {
 	c := rp.poollist
 
-	if existed, err := c.Exists(oldsid).Result(); err != nil || existed == 0 {
+	if existed, err := c.Exists(ctx, oldsid).Result(); err != nil || existed == 0 {
 		// oldsid doesn't exists, set the new sid directly
 		// ignore error here, since if it return error
 		// the existed value will be 0
-		c.Set(sid, "", time.Duration(rp.maxlifetime)*time.Second)
+		c.Set(ctx, sid, "", time.Duration(rp.maxlifetime)*time.Second)
 	} else {
-		c.Rename(oldsid, sid)
-		c.Expire(sid, time.Duration(rp.maxlifetime)*time.Second)
+		c.Rename(ctx, oldsid, sid)
+		c.Expire(ctx, sid, time.Duration(rp.maxlifetime)*time.Second)
 	}
-	return rp.SessionRead(context.Background(), sid)
+	return rp.SessionRead(ctx, sid)
 }
 
 // SessionDestroy delete redis session by id
 func (rp *Provider) SessionDestroy(ctx context.Context, sid string) error {
 	c := rp.poollist
-	c.Del(sid)
+	c.Del(ctx, sid)
 	return nil
 }
 
